requirements.txt
python-telegram-bot==21.4
anthropic==0.28.0
openai==1.30.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
pandas==2.2.2
openpyxl==3.1.2
pypdf==4.2.0
httpx==0.27.0